✨현재 로그인한 사용자 정보를 프로필에 표시, 로그아웃 버튼 구현
- 프로필 영역 컴포넌트 분리
